implement MBC1
Squashed commit of the following:

commit 532a78d85e5388dc0708cf361ea42118c41f11c2
Author: bmoxb <maxoblack@yahoo.com>
Date:   Fri Jul 28 19:13:49 2023 +0200

    allocate the correct amount of MBC1 RAM based on cartridge header

commit bac7ed6c2dbf86169573ab99580225445695e8e8
Author: bmoxb <maxoblack@yahoo.com>
Date:   Thu Jul 27 22:38:25 2023 +0200

    correct implementation of advanced banking

commit 72ec6d6bc8bc69eff1704d634a62f8f11de4ee4d
Author: bmoxb <maxoblack@yahoo.com>
Date:   Thu Jul 27 21:59:06 2023 +0200

    implement remaining MBC1 features

commit 426be80f8404afdb2177f7cb8ab9e3fc4b5c5fb5
Author: bmoxb <maxoblack@yahoo.com>
Date:   Thu Jul 27 21:02:56 2023 +0200

    write tests for advanced MBC1 features (currently failing)

commit 70220c889e1f479387304f31322e937aa93c8f49
Author: bmoxb <maxoblack@yahoo.com>
Date:   Thu Jul 13 20:02:21 2023 +0200

    mark MBC-1 as complete in README.md and add more games to list

commit c8f68ea3a41177db7a64b02b3d4c0b487aa5afce
Author: bmoxb <maxoblack@yahoo.com>
Date:   Mon Jul 10 23:58:02 2023 +0200

    add RAM to MBC1

commit 4b1c34c10496af5e44e91e3eb51d77e733150a54
Author: bmoxb <maxoblack@yahoo.com>
Date:   Sun Jul 9 19:39:39 2023 +0200

    (some) games using MBC1 now boot

+++ core/src/mbc/rom_only.rs
@@ -12,7 +12,7 @@ impl RomOnly {
 
 impl super::MemoryBankController for RomOnly {
     fn read8(&self, addr: u16) -> u8 {
-        self.cart.read8(addr)
+        self.cart.read8(addr as usize)
     }
 
     fn write8(&mut self, addr: u16, value: u8) {
